removing duplicate flake8 call and adding gui cavity tests
