+++ .github/workflows/python-app.yml
@@ -31,10 +31,8 @@ jobs:
           if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
       - name: Lint with flake8
         run: |
-          # stop the build if there are Python syntax errors or undefined names
-          flake8 . --count --select=E9,F63,F7,F82 --show-source --statistics
-          # exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
-          flake8 . --count --max-complexity=10 --max-line-length=120 --statistics
+          # stop the build if there are flake8 errors
+          flake8 . --count --max-complexity=10 --max-line-length=120 --show-source --statistics
       - name: Test with pytest
         run: |
           QT_QPA_PLATFORM=offscreen PYDM_DEFAULT_PROTOCOL=fake python -m pytest -rx
